pyproject.toml
[project]
name = "rosetta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.3",
    "google-cloud-discoveryengine>=0.17.0",
    "requests>=2.32.5",
]
